refactor: move to testhelper.terraform
